Implementation of the getl() command
getl() takes the same number of paramters as get(). On success
the item will be locked for a period of 15 seconds.
The current cas value of the item is also returned by membase
following a successful getl operation. This cas value is stored
in an internal hash table which is indexed by the memcache key.

Subsequently when a set() is called, if a cas value is present
for the referenced key, we will forward a cas operation
to the MC using the cas value that we looked up from the hash table.


git-svn-id: https://svn.zynga.com/svn/zcc/incubator/php-pecl-memcache-zynga/trunk@1733 93ba844f-b7fb-4add-93bc-7020b7bd55bf