Fixing the bugs in pecl-memcache
Following bugs have been fixed:
SEG-8839 (getmultybykey returns a null 'value' for odd flags): this issue
was happening since we were inserting value field in the return array with null
value and later we checking that if value is not inserted than we were inserting
the actual value.So the value was not gettig inserted. In the fix, we get the
content of value in test_Value and check the if it is is null, then we insert
into the return array.
SEG-8554 (pecl-memcache by default scans all membases for a particular key):
Set the default value of memcache.allow_failover to false.
SEG-8866 (getmultibykey losing refcount): The return value of getmultibykey was
beeing store in variable which was again used to store the return value of next
getmultibykey return value.So the php was freeing the variable because of which
refcount of shard key become zero and in next getmultibykey call, it became invalid
memory.In the fix, increase the refcount of shardkey.